chatConvoPublicById
