+++ package.json
@@ -22,6 +22,7 @@
         "build": "rm -rf dist && tsc --skipLibCheck && npm run client",
         "client": "node dist/scripts/copyClient.js",
         "test": "mocha src/*test.ts --exit",
+        "convo": "mocha src/chat/methods/ChatConvo.test.ts --exit",
         "test-convo": "mocha src/chat/methods/*test.ts --exit",
         "eslint": "eslint ./src --fix --ext=ts"
     },
